+++ src/spyglass/common/common_dandi.py
@@ -8,7 +8,10 @@ import h5py
 import pynwb
 from fsspec.implementations.cached import CachingFileSystem
 
-from spyglass.utils import logger
+from spyglass.common.common_usage import Export, ExportSelection
+from spyglass.settings import export_dir
+from spyglass.utils import SpyglassMixin, logger
+from spyglass.utils.sql_helper_fn import SQLDumpHelper
 
 try:
     import dandi.download
@@ -23,22 +26,15 @@ try:
 
 except (ImportError, ModuleNotFoundError) as e:
     (
-        dandi.download,
-        dandi.organize,
-        dandi.upload,
-        dandi.validate,
+        dandi,
         known_instances,
         DandiAPIClient,
         get_metadata,
         OrganizeInvalid,
         Severity,
-    ) = [None] * 9
+    ) = [None] * 6
     logger.warning(e)
 
-
-from spyglass.common.common_usage import Export
-from spyglass.settings import export_dir
-from spyglass.utils import SpyglassMixin, logger
 
 schema = dj.schema("common_dandi")
 
@@ -183,6 +179,26 @@ class DandiPath(SpyglassMixin, dj.Manual):
             for t in translations
         ]
         self.insert(translations, ignore_extra_fields=True)
+
+    def write_mysqldump(self, export_key: dict):
+        """Write a MySQL dump script to the paper directory for DandiPath."""
+        key = (Export & export_key).fetch1("KEY")
+        paper_id = (Export & key).fetch1("paper_id")
+        spyglass_version = (ExportSelection & key).fetch(
+            "spyglass_version", limit=1
+        )[0]
+
+        self.compare_versions(
+            spyglass_version,
+            msg="Must use same Spyglass version for export and Dandi",
+        )
+
+        sql_dump = SQLDumpHelper(
+            paper_id=paper_id,
+            docker_id=None,
+            spyglass_version=spyglass_version,
+        )
+        sql_dump.write_mysqldump(self & export_key, file_suffix="_dandi")
 
 
 def _get_metadata(path):
