+++ src/spyglass/utils/dj_mixin.py
@@ -15,10 +15,11 @@ from datajoint.logging import logger as dj_logger
 from datajoint.table import Table
 from datajoint.utils import get_master, to_camel_case, user_choice
 from networkx import NetworkXError
+from packaging.version import parse as version_parse
 from pymysql.err import DataError
 
 from spyglass.utils.database_settings import SHARED_MODULES
-from spyglass.utils.dj_helper_fn import (  # NonDaemonPool,
+from spyglass.utils.dj_helper_fn import (
     NonDaemonPool,
     fetch_nwb,
     get_nwb_table,
@@ -736,6 +737,30 @@ class SpyglassMixin:
         from spyglass import __version__ as sg_version
 
         return ".".join(sg_version.split(".")[:3])  # Major.Minor.Patch
+
+    def compare_versions(
+        self, version: str, other: str = None, msg: str = None
+    ) -> None:
+        """Compare two versions. Raise error if not equal.
+
+        Parameters
+        ----------
+        version : str
+            Version to compare.
+        other : str, optional
+            Other version to compare. Default None. Use self._spyglass_version.
+        msg : str, optional
+            Additional error message info. Default None.
+        """
+        if self._test_mode:
+            return
+
+        other = other or self._spyglass_version
+
+        if version_parse(version) != version_parse(other):
+            raise RuntimeError(
+                f"Found mismatched versions: {version} vs {other}\n{msg}"
+            )
 
     @cached_property
     def _export_table(self):
